feat: close #10 Seperate code from App.tsx, Add absolute path support for components directory
