+++ craco.config.js
@@ -1,0 +1,14 @@
+const { CracoAliasPlugin } = require("react-app-alias");
+
+module.exports = {
+  plugins: [
+    {
+      plugin: CracoAliasPlugin,
+      options: {
+        source: "tsconfig",
+        baseUrl: ".",
+        tsConfigPath: "./tsconfig.paths.json",
+      },
+    },
+  ],
+};
