+++ package.json
@@ -4,24 +4,27 @@
   "private": true,
   "dependencies": {
     "@babel/preset-typescript": "^7.23.2",
+    "@craco/craco": "^7.1.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^29.5.6",
+    "@types/react": "^18.2.30",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "openai": "^4.12.1",
     "react": "^18.2.0",
+    "react-app-alias": "^2.2.2",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "typescript": "^5.2.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "jest",
-    "eject": "react-scripts eject"
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "eject": "craco eject"
   },
   "eslintConfig": {
     "extends": [
@@ -42,6 +45,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-react": "^7.22.15",
     "jsdom": "^22.1.0"
